## IEEE bbc:microbit Training Repository
This repository is designed for both teachers and students on the bbc:microbit. Sample applications and training materials are designed mainly in Python for use on the microbit. Python is a fantastic language to learn how to program, and the bbc:microbit offers a very complete platform to learn both electronics and the Python launguage itself.

We will direct students to folders designed for them within this repository. We will also have folders for teachers and instructors of the on-line lessons. But there are no restrictions on this repository and the materials are welcome to all.

### References
<ul>

* bbc:Microbit Website
www.microbit.org

* Micropython Reference
https://microbit-micropython.readthedocs.io/en/v1.0.1/

* Mu-Editor Installation:
https://codewith.mu/en/download

* Seminar Slides:

* Python Language:
https://www.w3schools.com/python/default.asp
https://www.python.org/	(advanced full reference for Python)
https://wiki.python.org/moin/BeginnersGuide  (still pretty advanced)

* Electronics Basics Vendors
Adafruit Industries: www.adafruit.com
Sparkfun electronics: www.sparkfun.com

<li>